Cleanup

+++ app/src/main/java/com/model/shortcuts/ActionID.java
@@ -1,14 +1,14 @@
 package com.model.shortcuts;
 
 public class ActionID {
-    public final static int MOVE = (int) 1001;
-    public final static int ATTACK_1 = (int) 1002;
-    public final static int ATTACK_2 = (int) 1003;
-    public final static int BACK = (int) 1201;
-    public final static int NEXT_TURN = (int) 1200;
+    public final static int MOVE = 1001;
+    public final static int ATTACK_1 = 1002;
+    public final static int ATTACK_2 = 1003;
+    public final static int BACK = 1201;
+    public final static int NEXT_TURN = 1200;
 
-    public final static int NAME = (int) 90001;
-    public final static int HP = (int) 90002;
-    public final static int STEPS = (int) 90003;
-    public final static int CLASS = (int) 90004;
+    public final static int NAME = 90001;
+    public final static int HP = 90002;
+    public final static int STEPS = 90003;
+    public final static int CLASS = 90004;
 }
